move matrix_to_vector
